refactor: extract SectionHeader to lib/widgets/section_header.dart
Consolidate identical _SectionHeader implementations from
settings_screen, devices_screen, and notification_settings_screen
into a single shared public widget.

Closes #48
