+++ lib/widgets/section_header.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class SectionHeader extends StatelessWidget {
+  const SectionHeader({super.key, required this.label});
+  final String label;
+
+  @override
+  Widget build(BuildContext context) {
+    final cs = Theme.of(context).colorScheme;
+    final tt = Theme.of(context).textTheme;
+    return Padding(
+      padding: const EdgeInsets.only(left: 4, bottom: 8),
+      child: Text(
+        label,
+        style: tt.labelSmall?.copyWith(
+          color: cs.primary,
+          letterSpacing: 1.5,
+        ),
+      ),
+    );
+  }
+}
